move Cache to __init__.pyi
