+++ stubs/cachetools/cachetools/__init__.pyi
@@ -1,6 +1,24 @@
-from .cache import Cache as Cache
 from .decorators import cached as cached, cachedmethod as cachedmethod
 from .lfu import LFUCache as LFUCache
 from .lru import LRUCache as LRUCache
 from .rr import RRCache as RRCache
 from .ttl import TTLCache as TTLCache
+
+from typing import Callable, Generic, Iterator, TypeVar
+
+_KT = TypeVar("_KT")
+_VT = TypeVar("_VT")
+
+class Cache(MutableMapping[_KT, _VT], Generic[_KT, _VT]):
+    def __init__(self, maxsize: float, getsizeof: Callable[[_VT], float] | None = ...) -> None: ...
+    def __getitem__(self, key: _KT) -> _VT: ...
+    def __setitem__(self, key: _KT, value: _VT) -> None: ...
+    def __delitem__(self, key: _KT) -> None: ...
+    def __iter__(self) -> Iterator[_KT]: ...
+    def __len__(self) -> int: ...
+    @property
+    def maxsize(self) -> float: ...
+    @property
+    def currsize(self) -> float: ...
+    @staticmethod
+    def getsizeof(value: _VT) -> float: ...
